Changed how video caching works for faster 2nd through n startups. Removed some unused and unneeded values in the wrapper. Changed seeking to be percentage based instead of position based. Seeking when the video is not play now waits for the video to be playing (trying to seek while paused frequently fails)

+++ projects/video/gstreamer/src/ds/ui/sprite/video.h
@@ -34,7 +34,7 @@ namespace ds {
 			void				play();
 			void				stop();
 			void				pause();
-			void				seek(float t);
+			void				seek(double t); // 0.0 to 1.0 value
 			double				duration();
 			double				currentTime();
 			bool				isPlaying();
